+++ packages/ui/color-picker/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+## 0.0.5
+
+### Patch Changes
+
+- [#355](https://github.com/wpmudev/sui-react/pull/355)
+  [`503754f`](https://github.com/wpmudev/sui-react/commit/503754fe4f0d7745369493f06311bbc55d9322f6)
+  Thanks [@creador-dev](https://github.com/creador-dev)! - Popover popup
+  automatic placement.
+
+- [#355](https://github.com/wpmudev/sui-react/pull/355)
+  [`c37ddcb`](https://github.com/wpmudev/sui-react/commit/c37ddcb96d23a3af499de507cec2100b75264c4e)
+  Thanks [@creador-dev](https://github.com/creador-dev)! - Setup banner and
+  navigation custom icon prop
+
+- Updated dependencies
+  [[`503754f`](https://github.com/wpmudev/sui-react/commit/503754fe4f0d7745369493f06311bbc55d9322f6),
+  [`c37ddcb`](https://github.com/wpmudev/sui-react/commit/c37ddcb96d23a3af499de507cec2100b75264c4e),
+  [`36ba29c`](https://github.com/wpmudev/sui-react/commit/36ba29c920888612cf431d0a419cbf8a42e9838e)]:
+  - @wpmudev/sui-button@0.1.3
+  - @wpmudev/sui-input@0.1.1
+
 ## 0.0.4
 
 ### Patch Changes
@@ -19,6 +40,11 @@
   component
 
 - Updated dependencies
+  [[`ef804ad`](https://github.com/wpmudev/sui-react/commit/ef804adbfb5d783e066897e4771bc201f704351b),
+  [`ef804ad`](https://github.com/wpmudev/sui-react/commit/ef804adbfb5d783e066897e4771bc201f704351b),
+  [`ef804ad`](https://github.com/wpmudev/sui-react/commit/ef804adbfb5d783e066897e4771bc201f704351b)]:
+  - @wpmudev/sui-input@0.1.0
+  - @wpmudev/sui-button@0.1.2
   [[`de4695f`](https://github.com/wpmudev/sui-react/commit/de4695feb4886a3af66e5c4b51031915907e2490),
   [`a9ead37`](https://github.com/wpmudev/sui-react/commit/a9ead374fd02ebd63882d55bc01ab2fc7375732f)]:
   - @wpmudev/sui-button@0.1.2